+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## v1.1.0
+
+- add `swapf.swapPick` to open selected uris
+- stop search for alternatives if one pattern matches (overridable with force)
+- add c, cpp and h swap pattern
+- add cs, cs.designer swap pattern
+
 ## v1.0.1
 
 - add missing browser entry point